Refactor EdFiDataStandardListingGenerator to follow enhancer/generator pattern (#457)
* Initial plan

* Refactor EdFiDataStandardListingGenerator to follow enhancer/generator pattern


* Move extractDataType tests to DataCatalogEnhancer test file


---------
