+++ packages/metaed-plugin-edfi-data-catalog/src/index.ts
@@ -4,12 +4,13 @@
 // See the LICENSE and NOTICES files in the project root for more information.
 
 import { MetaEdPlugin } from '@edfi/metaed-core';
+import { enhance as dataCatalogEnhancer } from './enhancer/DataCatalogEnhancer';
 import { generate as dataStandardListingGenerator } from './generator/EdFiDataStandardListingGenerator';
 
 export function initialize(): MetaEdPlugin {
   return {
     validator: [],
-    enhancer: [],
+    enhancer: [dataCatalogEnhancer],
     generator: [dataStandardListingGenerator],
     shortName: 'edfiDataCatalog',
     configurationSchemas: new Map(),
